app/requirements_medical.txt
# Medical Image Encoder + Web App dependencies
# Install with: pip install -r app/requirements_medical.txt

# Web application
streamlit>=1.32.0

# Gemini LLM + embeddings (clinical simulation + case search)
google-genai>=1.0.0

# BiomedCLIP model loader (open_clip_torch, not open_clip)
open_clip_torch>=2.24.0

# Deep learning backend (CPU version — swap for torch-cuda if you have a GPU)
torch>=2.0.0
torchvision>=0.15.0

# Image loading
Pillow>=10.0.0

# Medical imaging datasets (no account required)
medmnist>=3.0.0

# Numerical ops (likely already installed via VectorAI DB client)
numpy>=1.24.0

# MultiCaRe dataset (real PubMed Central case reports)
multiversity
ipython

# Environment variable loading
python-dotenv

# Optional: DICOM file support (medical scanner files)
# Uncomment the line below if you need to process .dcm files
# pydicom>=2.4.0
